rm metagpt/utils/utils.py
