+++ metagpt/actions/action.py
@@ -18,8 +18,8 @@ from metagpt.llm import LLM
 from metagpt.logs import logger
 from metagpt.provider.base_gpt_api import BaseGPTAPI
 from metagpt.provider.postprecess.llm_output_postprecess import llm_output_postprecess
-from metagpt.utils.common import OutputParser
-from metagpt.utils.utils import general_after_log
+from metagpt.utils.common import OutputParser, general_after_log
+
 
 action_subclass_registry = {}
